requirements.txt
anyio==4.4.0
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-telegram-bot==21.3
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.30
typing_extensions==4.12.2
urllib3==2.2.1
